backend/requirements.txt
fastapi
uvicorn[standard]
numpy
elevenlabs
python-dotenv
fastrtc[vad]
soundfile
loguru
torch
supabase